+++ app/smartbrain_3000/web/_app/immutable/chunks/k5Gi4gQt.js
@@ -1,0 +1,1 @@
+import{Y as e}from"./C_H1zBAE.js";import{n as t}from"./9oWtiuEZ.js";var n=e({count:0});async function r(){try{n.count=(await t.unseenScheduleUpdates()).count}catch{}}export{n,r as t};